refactor: move body and header to its own file
